pyproject.toml
[project]
name = "diyu-agent"
version = "0.1.0"
description = "Diyu Agent - AI-powered intelligent work assistant platform"
# readme = "README.md"  # Uncomment when README exists
requires-python = ">=3.12"
license = { text = "PROPRIETARY" }
authors = [{ name = "Faye" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "pyjwt[crypto]>=2.10.0",
    "pyyaml>=6.0.2",
    "structlog>=24.4.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "jsonschema>=4.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "ipython>=8.30.0",
]
vector = [
    "pgvector>=0.3.6",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "T20",  # flake8-print (ban print in production)
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "S101",  # assert used in tests
    "S104",  # binding to all interfaces
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S603", "S607", "T20"]
"scripts/**" = ["T20"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests (no external deps)",
    "isolation: Isolation/RLS tests (needs DB)",
    "smoke: Smoke tests (fast subset)",
    "integration: Integration tests (needs services)",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@overload",
]
